+++ src/components/Dropdown/DropdownButton.jsx
@@ -22,31 +22,21 @@ const DropdownButton = ({
   type = "base",
   errorMessage = "",
 }) => {
-  const baseClasses = "flex w-auto min-h-9 items-center gap-x-1.5 cursor-pointer duration-200";
+  const baseClasses = "flex w-auto h-8 md:h-9 bg-black items-center gap-x-1.5 cursor-pointer duration-200";
 
   // 베이스 버튼(기능적 버튼)
   const baseButtonClass = `
-    ${baseClasses}justify-center text-gray-900
-    hover:text-gray-700
-    disabled:text-gray-400 disabled:cursor-not-allowed
+    ${baseClasses}justify-center text-gray-900 hover:text-gray-700 disabled:text-gray-400 disabled:cursor-not-allowed
   `;
 
   // 일반 버튼 UI
   const defaultButtonClass = `
-    ${baseClasses} border border-gray-300 rounded-md bg-white text-gray-900
-    hover:border-gray-500
-    focus:outline-2 focus:outline-gray-600 focus:outline-offset-[-2px]
-    disabled:bg-gray-100 disabled:border-gray-300 disabled:text-gray-400 disabled:cursor-not-allowed
+    ${baseClasses} border border-gray-300 rounded-md bg-white text-gray-900 hover:border-gray-500 focus:outline-2 focus:outline-gray-600 focus:outline-offset-[-2px] disabled:bg-gray-100 disabled:border-gray-300 disabled:text-gray-400 disabled:cursor-not-allowed
   `;
 
   // 셀렉트박스 스타일 버튼
   const selectButtonClass = `
-    ${baseClasses} w-full max-w-[320px] h-[50px] px-[16px] pr-[13px] justify-between
-    border rounded-[8px] border-gray-300 text-gray-600 text-base leading-6 tracking-[-1%]
-    hover:border-gray-600
-    focus:outline-2 focus:outline-gray-600 focus:outline-offset-[-2px]
-    active:border-2 active:border-gray-600 active:text-gray-900
-    disabled:pointer-events-none disabled:bg-gray-100 disabled:border-gray-300 disabled:text-gray-400 disabled:cursor-not-allowed
+    ${baseClasses} w-full max-w-[320px] h-[50px] px-[16px] pr-[13px] justify-between border rounded-[8px] border-gray-300 text-gray-600 text-base leading-6 tracking-[-1%] hover:border-gray-600 focus:outline-2 focus:outline-gray-600 focus:outline-offset-[-2px] active:border-2 active:border-gray-600 active:text-gray-900 disabled:pointer-events-none disabled:bg-gray-100 disabled:border-gray-300 disabled:text-gray-400 disabled:cursor-not-allowed
   `;
 
   // type에 따른 클래스 매핑
