+++ src/components/Dropdown/index.js
@@ -1,4 +1,4 @@
-export { default as Dropdown } from "./Dropdown";
+export { default as Dropdown } from "./index.jsx";
 export { default as DropdownButton } from "./DropdownButton";
 export { default as DropdownList } from "./DropdownList";
 export { default as DropdownItem } from "./DropdownItem";
